requirements.txt
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.2
Flask==2.3.2
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
oauthlib==3.2.2
outcome==1.2.0
Pillow==10.0.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pymongo==4.4.0
PyMsgBox==1.0.9
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytweening==1.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.6
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.1
trio-websocket==0.10.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
wsproto==1.2.0
